refactor(api): restructured project infrastructure to maximize growth [PW-81]
