--- api/src/main/java/com/michaelhyi/auth/AuthController.java
@@ -1,30 +1,0 @@
-package com.michaelhyi.auth;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("api/v1/auth")
-public class AuthController {
-    private final AuthService service;
-
-    public AuthController(AuthService service) {
-        this.service = service;
-    }
-
-    @PostMapping("{email}")
-    public ResponseEntity<String> authenticate(
-            @PathVariable("email") String email) {
-        return ResponseEntity.ok(service.authenticate(email));
-    }
-
-    @GetMapping("validate-token/{token}")
-    public ResponseEntity<Boolean> validateToken(
-            @PathVariable("token") String token) {
-        return ResponseEntity.ok(service.validateToken(token));
-    }
-}
