+++ api/src/main/java/com/michaelhyi/dao/PostRepository.java
@@ -1,0 +1,11 @@
+package com.michaelhyi.dao;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.michaelhyi.entity.Post;
+
+import jakarta.transaction.Transactional;
+
+@Transactional
+public interface PostRepository extends JpaRepository<Post, String> {
+}
