+++ api/src/main/java/com/michaelhyi/dao/UserRepository.java
@@ -1,0 +1,14 @@
+package com.michaelhyi.dao;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.michaelhyi.entity.User;
+
+import jakarta.transaction.Transactional;
+
+@Transactional
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
