+++ api/src/main/java/com/michaelhyi/dto/PostRequest.java
@@ -1,0 +1,7 @@
+package com.michaelhyi.dto;
+
+public record PostRequest(
+        String id,
+        String text
+) {
+}
