+++ api/src/main/java/com/michaelhyi/exception/S3Exception.java
@@ -1,0 +1,7 @@
+package com.michaelhyi.exception;
+
+public class S3Exception extends RuntimeException {
+    public S3Exception() {
+        super("S3 is unable complete that action.");
+    }
+}
