--- api/src/main/java/com/michaelhyi/post/PostDestructuredRequest.java
@@ -1,7 +1,0 @@
-package com.michaelhyi.post;
-
-public record PostDestructuredRequest(
-        String title,
-        String content
-) {
-}
