--- api/src/main/java/com/michaelhyi/post/PostRequest.java
@@ -1,7 +1,0 @@
-package com.michaelhyi.post;
-
-public record PostRequest(
-        String id,
-        String text
-) {
-}
