--- api/src/main/java/com/michaelhyi/s3/S3Buckets.java
@@ -1,18 +1,0 @@
-package com.michaelhyi.s3;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ConfigurationProperties(prefix = "aws.s3.buckets")
-public class S3Buckets {
-    private String blog;
-
-    public String getBlog() {
-        return blog;
-    }
-
-    public void setBlog(String blog) {
-        this.blog = blog;
-    }
-}
