--- api/src/main/java/com/michaelhyi/s3/S3Service.java
@@ -1,59 +1,0 @@
-package com.michaelhyi.s3;
-
-import java.io.IOException;
-
-import org.springframework.stereotype.Service;
-
-import software.amazon.awssdk.core.ResponseInputStream;
-import software.amazon.awssdk.core.sync.RequestBody;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
-import software.amazon.awssdk.services.s3.model.GetObjectRequest;
-import software.amazon.awssdk.services.s3.model.GetObjectResponse;
-import software.amazon.awssdk.services.s3.model.PutObjectRequest;
-
-@Service
-public class S3Service {
-    private final S3Client client;
-
-    public S3Service(S3Client client) {
-        this.client = client;
-    }
-
-    public void putObject(String bucketName, String key, byte[] file) {
-        PutObjectRequest putObjectRequest = PutObjectRequest
-                .builder()
-                .bucket(bucketName)
-                .key(key)
-                .build();
-
-        client.putObject(putObjectRequest, RequestBody.fromBytes(file));
-    }
-
-    public byte[] getObject(String bucketName, String key) {
-        GetObjectRequest getObjectRequest = GetObjectRequest
-                .builder()
-                .bucket(bucketName)
-                .key(key)
-                .build();
-
-        ResponseInputStream<GetObjectResponse> res =
-                client.getObject(getObjectRequest);
-
-        try {
-            return res.readAllBytes();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public void deleteObject(String bucketName, String key) {
-        DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest
-                .builder()
-                .bucket(bucketName)
-                .key(key)
-                .build();
-
-        client.deleteObject(deleteObjectRequest);
-    }
-}
