--- api/src/main/java/com/michaelhyi/security/ApplicationConfig.java
@@ -1,54 +1,0 @@
-package com.michaelhyi.security;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-import com.michaelhyi.user.UserRepository;
-
-@Configuration
-public class ApplicationConfig {
-    @Value("#{'${auth.whitelisted-emails}'.split(',')}")
-    private List<String> WHITELISTED_EMAILS;
-    private final UserRepository repository;
-
-    public ApplicationConfig(UserRepository repository) {
-        this.repository = repository;
-    }
-
-    @Bean
-    public AuthenticationProvider authenticationProvider() {
-        DaoAuthenticationProvider daoAuthenticationProvider =
-                new DaoAuthenticationProvider();
-        daoAuthenticationProvider.setUserDetailsService(userDetailsService());
-        return daoAuthenticationProvider;
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(
-            AuthenticationConfiguration authenticationConfiguration
-    ) throws Exception {
-        return authenticationConfiguration.getAuthenticationManager();
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        return username -> repository.findByEmail(username)
-                .orElseThrow(() ->
-                        new UsernameNotFoundException("User not found.")
-                );
-    }
-
-    @Bean
-    public List<String> whitelistedEmails() {
-        return WHITELISTED_EMAILS;
-    }
-}
