+++ api/src/main/java/com/michaelhyi/service/S3Service.java
@@ -1,0 +1,65 @@
+package com.michaelhyi.service;
+
+import java.io.IOException;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import com.michaelhyi.exception.S3Exception;
+
+import software.amazon.awssdk.core.ResponseInputStream;
+import software.amazon.awssdk.core.sync.RequestBody;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
+import software.amazon.awssdk.services.s3.model.GetObjectRequest;
+import software.amazon.awssdk.services.s3.model.GetObjectResponse;
+import software.amazon.awssdk.services.s3.model.PutObjectRequest;
+
+@Service
+public class S3Service {
+    @Value("${aws.s3.bucket}")
+    private String bucket;
+
+    private final S3Client client;
+
+    public S3Service(S3Client client) {
+        this.client = client;
+    }
+
+    public void putObject(String key, byte[] file) {
+        PutObjectRequest putObjectRequest = PutObjectRequest
+                .builder()
+                .bucket(bucket)
+                .key(key)
+                .build();
+
+        client.putObject(putObjectRequest, RequestBody.fromBytes(file));
+    }
+
+    public byte[] getObject(String key) {
+        GetObjectRequest getObjectRequest = GetObjectRequest
+                .builder()
+                .bucket(bucket)
+                .key(key)
+                .build();
+
+        ResponseInputStream<GetObjectResponse> res =
+                client.getObject(getObjectRequest);
+
+        try {
+            return res.readAllBytes();
+        } catch (IOException e) {
+            throw new S3Exception();
+        }
+    }
+
+    public void deleteObject(String key) {
+        DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest
+                .builder()
+                .bucket(bucket)
+                .key(key)
+                .build();
+
+        client.deleteObject(deleteObjectRequest);
+    }
+}
