--- api/src/main/java/com/michaelhyi/user/UserRepository.java
@@ -1,12 +1,0 @@
-package com.michaelhyi.user;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import jakarta.transaction.Transactional;
-
-@Transactional
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
