+++ api/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,57 +1,64 @@
-{"properties": [
-  {
-    "name": "auth.whitelisted-emails",
-    "type": "java.lang.String",
-    "description": "A description for 'auth.whitelisted-emails'"
-  },
-  {
-    "name": "aws.access-key",
-    "type": "java.lang.String",
-    "description": "A description for 'aws.access-key'"
-  },
-  {
-    "name": "aws.secret-key",
-    "type": "java.lang.String",
-    "description": "A description for 'aws.secret-key'"
-  },
-  {
-    "name": "aws.region",
-    "type": "java.lang.String",
-    "description": "A description for 'aws.region'"
-  },
-  {
-    "name": "aws.s3.buckets.blog",
-    "type": "java.lang.String",
-    "description": "A description for 'aws.s3.buckets.blog'"
-  },
-  {
-    "name": "security.jwt.secret-key",
-    "type": "java.lang.String",
-    "description": "A description for 'security.jwt.secret-key'"
-  },
-  {
-    "name": "security.jwt.expiration",
-    "type": "java.lang.String",
-    "description": "A description for 'security.jwt.expiration'"
-  },
-  {
-    "name": "security.cors.allowed-origins",
-    "type": "java.lang.String",
-    "description": "A description for 'security.cors.allowed-origins'"
-  },
-  {
-    "name": "security.cors.allowed-methods",
-    "type": "java.lang.String",
-    "description": "A description for 'security.cors.allowed-methods'"
-  },
-  {
-    "name": "security.cors.allowed-headers",
-    "type": "java.lang.String",
-    "description": "A description for 'security.cors.allowed-headers'"
-  },
-  {
-    "name": "security.cors.exposed-headers",
-    "type": "java.lang.String",
-    "description": "A description for 'security.cors.exposed-headers'"
-  }
-]}
+{
+  "properties": [
+    {
+      "name": "auth.whitelisted-emails",
+      "type": "java.lang.String",
+      "description": "A description for 'auth.whitelisted-emails'"
+    },
+    {
+      "name": "aws.access-key",
+      "type": "java.lang.String",
+      "description": "A description for 'aws.access-key'"
+    },
+    {
+      "name": "aws.secret-key",
+      "type": "java.lang.String",
+      "description": "A description for 'aws.secret-key'"
+    },
+    {
+      "name": "aws.region",
+      "type": "java.lang.String",
+      "description": "A description for 'aws.region'"
+    },
+    {
+      "name": "aws.s3.buckets.blog",
+      "type": "java.lang.String",
+      "description": "A description for 'aws.s3.buckets.blog'"
+    },
+    {
+      "name": "security.jwt.secret-key",
+      "type": "java.lang.String",
+      "description": "A description for 'security.jwt.secret-key'"
+    },
+    {
+      "name": "security.jwt.expiration",
+      "type": "java.lang.String",
+      "description": "A description for 'security.jwt.expiration'"
+    },
+    {
+      "name": "security.cors.allowed-origins",
+      "type": "java.lang.String",
+      "description": "A description for 'security.cors.allowed-origins'"
+    },
+    {
+      "name": "security.cors.allowed-methods",
+      "type": "java.lang.String",
+      "description": "A description for 'security.cors.allowed-methods'"
+    },
+    {
+      "name": "security.cors.allowed-headers",
+      "type": "java.lang.String",
+      "description": "A description for 'security.cors.allowed-headers'"
+    },
+    {
+      "name": "security.cors.exposed-headers",
+      "type": "java.lang.String",
+      "description": "A description for 'security.cors.exposed-headers'"
+    },
+    {
+      "name": "aws.s3.bucket",
+      "type": "java.lang.String",
+      "description": "A description for 'aws.s3.bucket'"
+    }
+  ]
+}
