+++ api/src/test/java/com/michaelhyi/unit/dao/UserRepositoryTest.java
@@ -1,0 +1,39 @@
+package com.michaelhyi.unit.dao;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+import com.michaelhyi.dao.UserRepository;
+import com.michaelhyi.entity.User;
+
+@DataJpaTest
+class UserRepositoryTest {
+    @Autowired
+    private UserRepository underTest;
+    private static final String email = "test@mail.com";
+
+    @AfterEach
+    void tearDown() {
+        underTest.deleteAll();
+    }
+
+    @Test
+    void findByEmail() {
+        User user = new User(email);
+        underTest.save(user);
+
+        assertTrue(underTest
+                .findByEmail(email)
+                .isPresent());
+
+        assertEquals(user, underTest
+                .findByEmail(email)
+                .get()
+        );
+    }
+}
